moves prompt form to component
